+++ server/src/shared/infra/http/controllers/users.e2e.spec.ts
@@ -4,7 +4,7 @@ import { Roles } from '@shared/constants';
 import request from 'supertest';
 
 import { AppModule } from '@/app.module';
-import { RegisterUser } from '@users/use-cases/registerUser/register-user';
+import { RegisterUser } from '@users/use-cases';
 
 describe('Users Controller', () => {
   const registerUser = {
